+++ Server/tfg-server-app/app/Http/Controllers/PuntuacionController.php
@@ -8,6 +8,7 @@ use App\Models\Juego;
 use App\Models\Puntuacion;
 use App\Models\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class PuntuacionController extends Controller
 {
@@ -16,9 +17,22 @@ class PuntuacionController extends Controller
      */
     public function index()
     {
-        $puntuaciones = Puntuacion::all()->sortByDesc(['updated_at', 'puntuacion']);
-        Puntuacion::all()->unique('user_id')->countBy();
-        return view('puntuaciones.index', compact('puntuaciones'));
+        $this->authorize('viewAny', Puntuacion::class);
+        $user = Auth::user();
+        $puntuaciones = [];
+
+        if($user->isAdmin()) {
+            $puntuaciones = Puntuacion::orderByDesc('updated_at', 'puntuacion')->paginate($this->paginatesNumber);
+        } else {
+            $puntuaciones = Puntuacion::where('user_id', $user->id)->orderByDesc('updated_at', 'puntuacion')->paginate($this->paginatesNumber);
+        }
+
+        $extraData = [
+            'createButton' => $user->isAdmin(),
+            'actionButtons' => $user->isAdmin()
+        ];
+
+        return view('puntuaciones.index', compact('puntuaciones', 'extraData'));
     }
 
     /**
@@ -26,6 +40,7 @@ class PuntuacionController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Puntuacion::class);
         $users = User::all();
         $juegos = Juego::all();
         return view('puntuaciones.create', compact('users', 'juegos'));
@@ -36,6 +51,7 @@ class PuntuacionController extends Controller
      */
     public function store(StorePuntuacionRequest $request)
     {
+        $this->authorize('create', Puntuacion::class);
         $validate = $request->validated();
         Puntuacion::create($validate);
         return redirect(route('puntuaciones.index'));
@@ -46,6 +62,7 @@ class PuntuacionController extends Controller
      */
     public function show(Puntuacion $puntuacion)
     {
+        $this->authorize('view', $puntuacion);
         $users = User::all();
         $juegos = Juego::all();
         return view('puntuaciones.show', compact('users', 'juegos', 'puntuacion'));
@@ -56,6 +73,7 @@ class PuntuacionController extends Controller
      */
     public function edit(Puntuacion $puntuacion)
     {
+        $this->authorize('update', $puntuacion);
         $users = User::all();
         $juegos = Juego::all();
         return view('puntuaciones.edit', compact('users', 'juegos', 'puntuacion'));
@@ -66,6 +84,7 @@ class PuntuacionController extends Controller
      */
     public function update(UpdatePuntuacionRequest $request, Puntuacion $puntuacion)
     {
+        $this->authorize('update', $puntuacion);
         $validate = $request->validated();
         $puntuacion->update($validate);
         return redirect(route('puntuaciones.index'));
@@ -76,6 +95,7 @@ class PuntuacionController extends Controller
      */
     public function destroy(Puntuacion $puntuacion)
     {
+        $this->authorize('delete', $puntuacion);
         $puntuacion->delete();
         return redirect(route('puntuaciones.index'));
     }
